+++ ide_image_scripts/Rakefile.rb
@@ -1,5 +1,0 @@
-require 'rake'
-require 'kitchen/rake_tasks'
-
-desc 'Run kitchen integration tests'
-Kitchen::RakeTasks.new